requirements.txt
gpxpy>=1.6.0
Pillow>=10.0.0
exifread>=3.0.0
watchdog>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
geopy>=2.4.0